week1_project/requirements.txt
openai
python-dotenv
gradio
fastapi
modal
pypdf
requests
pydantic
